A few fixes for boolean search
I have added a chunk of code at the end of the file, with a few additions for boolean search such as phrase search and proximity search, also a few changes on how it parses and evaluates a boolean query (with nested queries, a bit more complex). Also, for not operations precomputing all_docs might help with speed? I tested these only for the example queries that I have added in my code!